backend/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-migrate==2.5.3
flask-login==0.5.0
flask-bcrypt==0.7.1
wtforms==2.3.1
email-validator==1.1.1
flask-cors==3.0.10
psycopg2-binary==2.9.9
python-dotenv==1.0.1
